refactor the poll action to use the encapsulated CallFunctionOn func
