+++ call.go
@@ -2,6 +2,7 @@ package chromedp
 
 import (
 	"context"
+	"encoding/json"
 
 	"github.com/chromedp/cdproto/runtime"
 )
@@ -65,3 +66,23 @@ func CallFunctionOn(functionDeclaration string, res interface{}, opt CallOption,
 // CallOption is a function to modify the runtime.CallFunctionOnParams
 // to provide more information.
 type CallOption = func(params *runtime.CallFunctionOnParams) *runtime.CallFunctionOnParams
+
+// errAppender is to help accumulating the arguments and simplifying error checks.
+//
+// see https://blog.golang.org/errors-are-values
+type errAppender struct {
+	args []*runtime.CallArgument
+	err  error
+}
+
+// append method calls the json.Marshal method to marshal the value and appends it to the slice.
+// It records the first error for future reference.
+// As soon as an error occurs, the append method becomes a no-op but the error value is saved.
+func (ea *errAppender) append(v interface{}) {
+	if ea.err != nil {
+		return
+	}
+	var b []byte
+	b, ea.err = json.Marshal(v)
+	ea.args = append(ea.args, &runtime.CallArgument{Value: b})
+}
